requirements.txt
python-telegram-bot==20.7
spotipy==2.23.0
Flask==2.3.3
aiosqlite==0.19.0
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.2
scikit-learn==1.4.2
